Story#112965483 - skip removing whoises for reserved and blocked domains if domain exists

+++ app/models/blocked_domain.rb
@@ -48,6 +48,8 @@ class BlockedDomain < ActiveRecord::Base
   end
 
   def remove_data
+    next if Domain.where(name: name).any?
+
     Whois::Record.where(name: name).delete_all
   end
 end
